feat: Docs: Remove dead links and amp up the guide
This project is still linked in many static site lists
and stared by many people. It's old and has it's quirks
(see https://stitcher.io/blog/a-simple-approach-to-static-generation),
but it's working fine and still easier to set up than
many competing projects.

So a little upgrade to the docs - removing dead links,
add sidenotes, explain some files - is probably okay.